backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain
faiss-cpu
pdfminer.six
python-docx
psycopg2-binary
sqlalchemy
httpx
requests
python-multipart
aiofiles
tqdm
whisper
textract
azure-cognitiveservices-vision-computervision
azure-identity
azure-storage-blob
